# api/resource/swagger/FeatureManager-swagger.json
{
    "swagger": "2.0",
    "paths": {
...
        "/users/{key}": {
            "delete": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "User"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    }
                ],
                "tags": [
                    "User"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDto"
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/UserRequestDto"
                        }
                    }
                ],
                "tags": [
                    "User"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDto"
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "name": "key",
                        "type": "string",
                        "in": "path",
                        "required": true,
                        "description": null
                    },
                    {
                        "name": "body",
                        "type": "object",
                        "in": "body",
                        "required": true,
                        "description": null,
                        "schema": {
                            "$ref": "#/definitions/UserRequestDto"
                        }
                    }
                ],
                "tags": [
                    "User"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "UserResponseDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "object",
                    "example": null
                },
                "key": {
                    "type": "object",
                    "example": null
                },
                "username": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "email",
                "key",
                "username"
            ]
        },
        "UserRequestDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "object",
                    "example": null
                },
                "password": {
                    "type": "object",
                    "example": null
                },
                "username": {
                    "type": "object",
                    "example": null
                }
            },
            "required": [
                "email",
                "password",
                "username"
            ]
        },
...
    },
...
    "host": "localhost:5030",
    "schemes": [
        "http"
    ],
    "basePath": "/feature-dataset-manager-api",
    "info": {
        "title": "FeatureManager",
        "description": "This is an api that manage samples and its features. The main goal is to retrieve the best sample match given a set of features",
        "version": "0.0.1",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Samuel Jansen",
            "email": "samuel.jansenn@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 / MIT License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    }
}